custom_nodes/doodle-tiles-node/requirements.txt
numpy==1.26.4
Pillow==10.3.0
scikit-image==0.22.0